+++ sat/cli/bootprep/input/image.py
@@ -25,6 +25,7 @@
 Defines class for images defined in the input file.
 """
 from abc import ABC, abstractmethod
+import fnmatch
 import logging
 import warnings
 
@@ -640,9 +641,16 @@ class ProductInputImage(BaseInputImage):
         return self.image_data['base']['product'].get('version')
 
     @property
+    @jinja_rendered
     def filter_prefix(self):
         """str or None: the filter prefix to use for filtering images or recipes from the product"""
         return get_val_by_path(self.image_data, 'base.product.filter.prefix')
+
+    @property
+    @jinja_rendered
+    def filter_wildcard(self):
+        """str or None: the filter pattern to use for filtering images or recipes from the product"""
+        return get_val_by_path(self.image_data, 'base.product.filter.wildcard')
 
     def filter_func(self, name):
         """Filter the given name.
@@ -653,7 +661,15 @@ class ProductInputImage(BaseInputImage):
         Returns:
             True if the filter matches the given name or there is no filter, False otherwise
         """
-        return self.filter_prefix is None or name.startswith(self.filter_prefix)
+        # Schema validation guarantees that there cannot be both a prefix and a wildcard,
+        # so it is possible to just check each individually and not worry about the case
+        # where both are present.
+        if self.filter_prefix:
+            return name.startswith(self.filter_prefix)
+        if self.filter_wildcard is not None:
+            return fnmatch.fnmatch(name, self.filter_wildcard)
+
+        return True
 
     @property
     def unfiltered_base_description(self):